// src/app/modules/hr/user/models/user-search.viewmodel.ts
export class UserSearchViewModel {
    id: string;
    fullName: string;
    userName: string;
    avatar: string;
    status: number;
    titleId: string;
    titleName: string;
    positionId: string;
    positionName: string;
    officeId: number;
    officeName: string;
    officeIdPath: string;
    joinedDate: Date;
    phoneNumber: string;
    statusName: string;
}