// src/app/modules/hr/user/models/user-detail.viewmodel.ts
import {UserTranslation} from './user-translation.model';
import {UserContact} from './user-contact.model';

export class UserDetailViewModel {
    id: string;
    fullName: string;
    userName: string;
    avatar: string;
    birthday?: Date;
    idCardNumber: string;
    idCardDateOfIssue?: Date;
    gender: number;
    ethnic: number;
    ethnicName: string;
    denomination: number;
    tin: string;
    bankingNumber: string;
    provinceId: number;
    provinceName: string;
    nationalId: number;
    districtId: number;
    districtName: string;
    marriedStatus: number;
    userStatus: number;
    officeId: number;
    officeIdPath: string;
    officeName: string;
    titleId: string;
    userType: number;
    passportId: string;
    passportDateOfIssue: Date;
    cardNumber: string;
    ext: string;
    enrollNumber?: number;
    academicRank: number;
    concurrencyStamp: string;
    joinedDate?: Date;
    userTranslations: UserTranslation[];
    userContacts: UserContact[];
}